More accessible report embed conversion (#19)
* Rename commands create -> submit; submit -> queue

"create" remains an alias of the same command

* Improve usability of report command

Able to use command on reports outside of #repro-queue now. Looks first for the message in the current channel. Can also supply a channel ID in the format `channelID-messageID` (holding shift on a message to expand the quickbar and clicking the ID button currently results in this channel-message format being copied, as does shift clicking Copy ID in the message context menu)

* Improve edit command usability

Same access improvements as with the report command

* Add support for stored Linux info

Was an oversight during original implementation

* Fix bugs in report & edit improvements